refactored file sturcture, to match cookiecutter data science project suggestions
