+++ docs/brainstorm.minder
