+++ src/data/crop_images.py
